requirements.txt
requests~=2.25.1
click8~=8.0.1
email-validator~=1.1.2
